requirements.txt
flask
python-dotenv
requests
authlib>=0.14.1
Flask-WTF
redis